Ordenando los componentes
