+++ src/app/app.component.html
@@ -2,7 +2,6 @@
 <main>
   <router-outlet></router-outlet>
 </main>
-
+<app-login></app-login>
 <app-contacto></app-contacto>
-
 <app-footer></app-footer>
