+++ src/app/app.component.ts
@@ -43,5 +43,4 @@ export class AppComponent {
         }
       });
   }
-
 }
