+++ src/app/app.module.ts
@@ -9,11 +9,13 @@ import { AppComponent } from './app.component';
 import { HttpClientModule } from '@angular/common/http';
 import { SharedModule } from '@shared/shared.module';
 import { CommonModule } from '@angular/common';
+import { HomeModule } from '@home/home.module';
 
 @NgModule({
   declarations: [
     AppComponent
   ],
+
   imports: [
     BrowserModule,
     CommonModule,
