+++ src/app/core/components/drag/drag.component.html
@@ -1,0 +1,1 @@
+<p>drag works!</p>
