+++ src/app/core/components/drag/drag.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-drag',
+  templateUrl: './drag.component.html',
+  styleUrl: './drag.component.scss'
+})
+export class DragComponent {
+
+}
