+++ src/app/core/components/loader/loader.component.ts
@@ -2,8 +2,7 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-loader',
-  standalone: true,
-  imports: [],
+  standalone: false,
   templateUrl: './loader.component.html',
   styleUrl: './loader.component.scss'
 })
