+++ src/app/core/components/login/login.component.html
@@ -1,0 +1,41 @@
+<section class="vh-90">
+  <div class="container py-1 h-100">
+    <div class="row d-flex justify-content-center align-items-center h-100">
+      <div class="col col-xl-10">
+        <div class="card" style="border-radius: 1rem;">
+          <div class="row g-0">
+            <div class="col-md-6 col-lg-5 d-none d-md-block">
+              <img src="assets/img/pexels-anna-shvets-3727461.png" alt="login form" class="img-fluid"
+                style="border-radius: 1rem 0 0 1rem;" />
+            </div>
+            <div class="col-md-6 col-lg-7 d-flex align-items-center bg-dark">
+              <div class="card-body p-4 p-lg-5 text-black">
+                <form>
+                  <div class="d-flex align-items-center mb-3 pb-2">
+                    <span class="h1 fw-bold mb-0" style="color: rgb(255, 106, 0);">Entrevistador</span>
+                  </div>
+                  <h5 class="fw-bold mb-3 pb-3 text-white " style="letter-spacing: 2px;">Ingresa a tu cuenta</h5>
+                  <div class="form-outline mb-2">
+                    <label class="form-label text-white" for="form2Example17">Correo eléctronico</label>
+                    <input type="email" id="form2Example17" class="form-control form-control-lg mb-2" />
+                  </div>
+                  <div class="form-outline mb-2 mt-2">
+                    <label class="form-label text-white" for="form2Example27">Contraseña</label>
+                    <input type="password" id="form2Example27" class="form-control form-control-lg mb-2" />
+                  </div>
+                  <div class="pt-4 mb-2 mt-2">
+                    <button class="btn btn-outline-primary btn-lg btn-block" type="button"
+                      (click)="login">Login</button>
+                  </div>
+                  <a class="small text-muted text-white" href="#!">Forgot password?</a>
+                  <p class="mb-5 pb-lg-2 text-white">¿No tienes cuenta? <a routerLink="/registro"
+                      style="color: #ff6200; text-decoration: none;">Registrate aquí</a></p>
+                </form>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</section>
