+++ src/app/core/components/perfil/perfil.component.html
@@ -1,0 +1,1 @@
+<p>perfil works!</p>
