+++ src/app/core/components/perfil/perfil.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-perfil',
+  templateUrl: './perfil.component.html',
+  styleUrl: './perfil.component.scss'
+})
+export class PerfilComponent {
+
+}
