+++ src/app/core/components/registro/registro.component.html
@@ -1,0 +1,1 @@
+<p>registro works!</p>
