+++ src/app/core/components/registro/registro.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-registro',
+  templateUrl: './registro.component.html',
+  styleUrl: './registro.component.scss'
+})
+export class RegistroComponent {
+
+}
