+++ src/app/feature/home/home-login/home-login.component.html
@@ -1,30 +1,5 @@
 <h1 class="text-center fw-bold my-4">Bienvenido/a Logeado</h1>
 
-<!-- CARRUSEL DE IMAGENES -->
-
-<div id="carouselExampleFade" class="carousel slide carousel-fade border-rounded">
-  <div class="carousel-inner">
-    <div class="carousel-item active">
-      <img src="../../assets/img/istockphoto.jpg" class="d-block img-custom" alt="...">
-    </div>
-    <div class="carousel-item">
-      <img src="../../assets/img/istockphoto-1.jpg" class="d-block img-custom" alt="...">
-    </div>
-    <div class="carousel-item">
-      <img src="../../assets/img/istockphoto-2.jpg" class="d-block img-custom" alt="...">
-    </div>
-  </div>
-
-  <button class="carousel-control-prev" type="button" data-bs-target="#carouselExampleFade" data-bs-slide="prev">
-    <span class="carousel-control-prev-icon" aria-hidden="true"></span>
-    <span class="visually-hidden">Previous</span>
-  </button>
-  <button class="carousel-control-next" type="button" data-bs-target="#carouselExampleFade" data-bs-slide="next">
-    <span class="carousel-control-next-icon" aria-hidden="true"></span>
-    <span class="visually-hidden">Next</span>
-  </button>
-</div>
-
 <!-- PREGUNTAS PREDETERMINADAS SIN LOGUEARSE -->
 <section id="preguntas">
   <div class="table-responsive">
@@ -33,7 +8,6 @@
         <thead>
           <tr>
             <th class="text-center fw-bold">Pregunta</th>
-            <th class="text-center fw-bold">Acción</th>
           </tr>
         </thead>
         <tbody class="container">
@@ -51,4 +25,8 @@
       </table>
     </div>
   </div>
+</section>
+
+<section id="contacto">
+  <app-contacto></app-contacto>
 </section>
