--- src/app/feature/home/home-login/home-login.component.scss
@@ -1,20 +1,0 @@
-.carousel {
-    height: 800px;
-    width: 74%;
-    margin: auto;
-  
-    .carousel-inner {
-      height: 100%;
-  
-      .carousel-item {
-        width: 100%;
-  
-        .img-custom {
-          width: 100%;
-          height: 100%;
-          object-fit: cover;
-        }
-      }
-    }
-  }
-
