+++ src/app/feature/home/home-sin-login/home-sin-login.component.html
@@ -73,3 +73,7 @@
     </div>
   </div>
 </section>
+
+<section id="contacto">
+  <app-contacto></app-contacto>
+</section>
