+++ src/app/feature/home/home.component.ts
@@ -7,8 +7,6 @@ import { Component, Input } from '@angular/core';
 })
 
 export class HomeComponent {
-
   isLogged= true;
   username= "string";
-
 }
