--- src/app/shared/Components/registro/registro.component.html
@@ -1,1 +1,0 @@
-<p>registro works!</p>
