--- src/app/shared/Components/registro/registro.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-registro',
-  templateUrl: './registro.component.html',
-  styleUrl: './registro.component.scss'
-})
-export class RegistroComponent {
-
-}
