--- src/app/shared/components/contacto/contacto.component.html
@@ -1,45 +1,0 @@
-<form action="">
-  <div class="container contact mt-4 mb-4 bg-dark">
-    <div class="row">
-      <div class="col-md-3">
-        <div class="contact-info text-white mx-auto py-4">
-          <h2 class="text-center">Contactanos</h2>
-          <h4 class="text-center">¡Nos encantaría saber tu opinión!</h4>
-        </div>
-      </div>
-      <div class="col-md-8">
-        <div class="contact-form pt-4">
-          <div class="form-group">
-            <label class="control-label col-sm-2" for="fname">Primer nombre:</label>
-            <div class="col-sm-9">
-            <input type="text" class="form-control" id="fname" placeholder="Escribe tu primer nombre" name="fname">
-            </div>
-          </div>
-          <div class="form-group">
-            <label class="control-label col-sm-2" for="lname">Apellido:</label>
-            <div class="col-sm-9">
-            <input type="text" class="form-control" id="lname" placeholder="Escribe tu apellido" name="lname">
-            </div>
-          </div>
-          <div class="form-group">
-            <label class="control-label col-sm-2" for="email">Correo:</label>
-            <div class="col-sm-9">
-            <input type="email" class="form-control" id="email" placeholder="Escribe tu correo" name="email">
-            </div>
-          </div>
-          <div class="form-group">
-            <label class="control-label col-sm-2" for="comment">Comentario:</label>
-            <div class="col-sm-9">
-            <textarea class="form-control" rows="5" id="comment"></textarea>
-            </div>
-          </div>
-          <div class="form-group">
-            <div class="col-sm-offset-2 col-sm-10">
-            <button type="submit" class="btn btn-outline-primary">Enviar</button>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</form>
