--- src/app/shared/components/contacto/contacto.component.scss
@@ -1,22 +1,0 @@
-.container {
-  background-color: rgb(30, 30, 37) !important;
-  border-top-left-radius: 40px;
-  border-bottom-right-radius: 40px;
-
-  .contact-info {
-    text-align: center;
-    color: white;
-    margin-top: 8%;
-  }
-
-  label {
-    color: white;
-    margin-bottom: 0;
-  }
-
-  .btn {
-    margin: 10px 0;
-    font-weight: 600;
-    padding: 6px 40px;
-  }
-}
