--- src/app/shared/components/contacto/contacto.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-contacto',
-  templateUrl: './contacto.component.html',
-  styleUrl: './contacto.component.scss'
-})
-export class ContactoComponent {
-
-}
