+++ src/app/shared/components/footer/footer.component.html
@@ -6,17 +6,17 @@
           <h6 class="text-uppercase mb-2 font-weight-bold text-center">
             Entrevistador Inteligente
           </h6>
-          <div class="image-container"> <img class="mx-auto" src="assets/img/pregunta.png" alt="Logo">
+          <div class="image-container">
+            <img class="mx-auto" src="assets/img/pregunta.png" alt="Logo">
           </div>
-          <p class="mt-2  text-center">
+          <p class="mt-2 text-center">
             En Entrevistador Inteligente, te preparamos para entrevistas reales con apoyo de inteligencia artificial.
-            <br>
-            ¡Empieza hoy mismo!
           </p>
+          <p class="text-center fw-bold">¡Empieza hoy mismo!</p>
         </div>
         <hr class="w-100 clearfix d-md-none" />
         <div class="col-md-2 col-lg-2 col-xl-2 mx-auto mt-3">
-          <h6 class="text-uppercase mb-4 font-weight-bold text-center text-white" >Servicios</h6>
+          <h6 class="text-uppercase mb-4 font-weight-bold text-center">Servicios</h6>
           <ul class="navbar-nav justify-content-center flex-grow-1 pe-1 mt-0">
             <li class="nav-item">
               <a class="nav-link mx-lg-1 fw-bold" [ngClass]="{'active': activo === 'Entrevistas simuladas'}"
@@ -63,7 +63,7 @@
     </section>
   </div>
   <div class="text-center p-3" style="background-color: rgba(0, 0, 0, 0.2)">
-    © 2024 Copyright:
+    © 2024 Copyright
     <a class="text-white" style="text-decoration: none;">Entrevistador inteligente</a>
   </div>
 </footer>
