+++ src/app/shared/components/footer/footer.component.scss
@@ -1,18 +1,33 @@
 .footer {
-  background-color: rgb(30, 30, 37);
+  background-color: #212529;
   bottom: 0;
 
   .image-container {
     text-align: center;
 
     img.mx-auto {
-      height: 60px;
+      height: 50px;
       margin: 0 auto;
     }
   }
 
+  h6 {
+    color: white;
+    font-size: 16px;
+  }
+
   .navbar-nav > li {
     list-style: none;
+
+    .nav-item {
+      margin: -2px;
+    }
+
+  }
+
+  hr {
+    margin-top: 4px;
+    width: 94% !important;
   }
 
   .nav-link,
