+++ src/app/shared/shared.module.ts
@@ -1,20 +1,18 @@
 // MODULOS
 import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { HomeModule } from '@home/home.module';
 import { RouterModule } from '@angular/router';
 import { routes } from '../app.routes';
+
+// COMPONENTES
 import { MenuComponent } from './components/menu/menu.component';
 import { FooterComponent } from './components/footer/footer.component';
-import { ContactoComponent } from './components/contacto/contacto.component';
-import { LoginComponent } from './components/login/login.component';
+import { HomeModule } from '@home/home.module';
 
 @NgModule({
   declarations: [
     MenuComponent,
     FooterComponent,
-    ContactoComponent,
-    LoginComponent
   ],
   imports: [
     CommonModule,
@@ -23,9 +21,7 @@ import { LoginComponent } from './components/login/login.component';
   ],
   exports: [
     MenuComponent,
-    ContactoComponent,
-    FooterComponent,
-    LoginComponent
+    FooterComponent
   ],
   providers: [],
 })
